backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23

passlib[bcrypt]==1.7.4
python-multipart==0.0.6

pydantic==2.5.0
pydantic-settings==2.1.0

langchain==0.1.16
langchain-community>=0.0.32,<0.1
langchain-openai>=0.0.8,<0.1
openai>=1.10.0,<2.0.0
faiss-cpu==1.7.4
numpy==1.26.4
pandas==2.1.4

python-dotenv==1.0.0
alembic==1.12.1